+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add support for Deck or Slide-level transitions
 - Add default transitions for Fade, Slide, and None
 - Fixes Full Screen component for Chrome/FF, adds support for Safari
+- Added support for dual-browser tab mode for presenter mode in all browsers
 
 ## 6.0.0-alpha.7